Split up App into separate tabs for budget, income, and spending
